remove CellInputRendereContext
