+++ packages/iris-grid/src/IrisGridContextProvider.tsx
@@ -1,13 +1,41 @@
+/* eslint-disable react-refresh/only-export-components -- exports non-component values alongside the provider */
+import React, { createContext, type ReactNode, useMemo } from 'react';
 import { useTheme } from '@deephaven/components';
-import { createContext, type ReactNode, useContext, useMemo } from 'react';
-import { type CellInputRendererRegistry } from '@deephaven/grid';
-import CellInputRendererContext, {
-  DEFAULT_REGISTRY,
-} from './CellInputRendererContext';
+import {
+  type CellInputRendererFn,
+  type CellInputRendererRegistry,
+  type CellInputProps,
+} from '@deephaven/grid';
+import CellDropdownField from './CellDropdownField';
 import {
   createDefaultIrisGridTheme,
   type IrisGridThemeType,
 } from './IrisGridTheme';
+import {
+  STRING_LIST_RESTRICTION_TYPE,
+  type StringListRestriction,
+} from './IrisGridModel';
+
+export type { CellInputRendererFn, CellInputRendererRegistry, CellInputProps };
+
+const renderStringListRestriction: CellInputRendererFn = (
+  props: CellInputProps
+): ReactNode => {
+  const { restrictions } = props;
+  const { allowedValues } = restrictions[0] as StringListRestriction;
+  // eslint-disable-next-line react/jsx-props-no-spreading
+  return <CellDropdownField {...props} options={allowedValues} />;
+};
+renderStringListRestriction.preservesExistingValue = true;
+
+/**
+ * The default registry containing DHC's built-in cell input renderers.
+ * Pass this (or a superset of it) as the `registry` prop of
+ * {@link IrisGridContextProvider} to include the DHC defaults.
+ */
+export const DEFAULT_REGISTRY: CellInputRendererRegistry = new Map([
+  [STRING_LIST_RESTRICTION_TYPE, renderStringListRestriction],
+]);
 
 /**
  * The context value for the IrisGridContextProvider.
@@ -34,11 +62,14 @@ export interface IrisGridContextProviderProps {
   children: ReactNode;
   /* The density of the grid. Defaults to regular */
   density?: 'compact' | 'regular' | 'spacious';
+  /** Registry of cell input renderer functions. Defaults to the DHC built-in renderers. */
+  registry?: CellInputRendererRegistry;
 }
 
 export function IrisGridContextProvider({
   children,
   density = 'regular',
+  registry = DEFAULT_REGISTRY,
 }: IrisGridContextProviderProps): JSX.Element {
   const { activeThemes } = useTheme();
 
@@ -49,15 +80,13 @@ export function IrisGridContextProvider({
     [activeThemes]
   );
 
-  const cellInputRendererRegistry = useContext(CellInputRendererContext);
-
   const contextValue = useMemo(
     () => ({
       theme: gridTheme,
       density,
-      cellInputRendererRegistry,
+      cellInputRendererRegistry: registry,
     }),
-    [gridTheme, density, cellInputRendererRegistry]
+    [gridTheme, density, registry]
   );
 
   return (
